Renamed friendsFinder to friends
