+++ app.js
@@ -34,7 +34,7 @@ app.use('/', require('./routes/index'));
 app.use('/users', require('./routes/users'));
 app.use('/register', require('./routes/register'));
 app.use('/chatroom', require('./routes/chatroom'));
-app.use('/friendfinder', require('./routes/friendfinder'));
+app.use('/friends', require('./routes/friends'));
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
